refactor: SSR in register page
